Serverless Framework setup
#TODO 1.Fix the token_required decorator function to work with the event arg instead of flask.request

#TODO 2.Test the functions with the new endpoint at aws, add permissions for DynamoDB and S3 into serverless.yaml if needed
